tests/wpt/tests/html/dom/render-blocking/element-render-blocking-002.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Frames starts after href element is parsed before the end</title>

<link rel=expect href="#third" blocking="render">
<script>
async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("third")));
    t.step(() => assert_false(!!document.getElementById("last")));
    t.done();
  });
}, "blocking defers until needed element is parsed");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="third"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="fourth"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
